pyproject.toml
[tool.poetry]
name = "telegram-legal-bot"
version = "0.1.0"
description = "Telegram bot for legal consultations powered by OpenAI GPT"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    # Устанавливаем модуль с точкой входа по layout src/
    { include = "telegram_legal_bot", from = "src" },
    # Параллельно публикуем текущий модуль "src" (используется в импортах: src.core)
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.12"

# Core dependencies
aiogram = ">=3.4.0,<4.0.0"
openai = "^1.40"
httpx = {version = "^0.27", extras = ["http2"]}
python-dotenv = "^1.0.1"
aiohttp-socks = "^0.10.1"
requests = "^2.32.5"
PyPDF2 = "^3.0.1"
pdfplumber = "^0.11.4"
python-docx = "^1.1.0"
pytesseract = "^0.3.10"
pdf2image = "^1.17.0"
Pillow = "^10.4.0"
reportlab = "^4.2.0"
textract = {version = "^1.6.5", optional = true}
# Advanced database support
aiosqlite = "^0.20.0"

# Redis for caching and scaling
redis = {version = "^5.0.6", optional = true}

# Monitoring and metrics
prometheus-client = {version = "^0.20.0", optional = true}

# System monitoring
psutil = {version = "^5.9.0", optional = true}

# Data validation and serialization
pydantic = {version = "^2.5.0", optional = true}

[tool.poetry.extras]
prod = ["redis", "prometheus-client", "psutil", "pydantic", "textract"]
monitoring = ["prometheus-client", "psutil"]
scaling = ["redis", "pydantic"]
full = ["redis", "prometheus-client", "psutil", "pydantic", "textract"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.10.0"
types-redis = "^4.6.0.20240425"

[tool.poetry.scripts]
telegram-legal-bot = "telegram_legal_bot.main:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

# --- стиль кода (по желанию, но удобно) ---
[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E","F","I","W","UP","N","B","PL"]
ignore = ["E501"]
